Introduce `DisConfig` and Add devWhitelisting
DisConfig is intended to be a static class that we can put
all of our configuration related options into and have the freedom
to access them in the code where-ever we want.

The devWhitelisting feature allows a dev to put their discord user id
in the config file after the bot token to ensure that their bot instance
only responds to them.

The above, coupled with !suppress or !blacklist should allow a dev to
only interact with their instance of the bot, while a live public
version runs on the same token.