Change project-jdk-name to 13.0.2
it might be a decent idea to gitignore .idea/misc.xml but I don't actually know. We ought to talk about this if further changes to this file break setups. Please try to not edit this file if you're able.

This may require you to change some gradle compiler settings in idea. If you'd like some help with that ping me on discord.