Fixed Opus bug in fatJar task
Fixed the bug in the fatJar task that caused opus to not be able to be included. This happened because the fatJar code only looked if something was a directory and assumed if it wasn't a directory, it must be a jar. Opus includes a .pom file, which caused the bug when the fatJar task tried to unzip it.