Correct Blacklist functionality  and command
When this was changed to use a switch statement instead of an if-else
chain, the blacklist check was moved to *after* all commands were
processed. This essentially meant that the feature was broken because of
said change.

This is now corrected so that the only usable command while blacklisted
is the `!blacklist` command, and the blacklist check is now gating the
try-catch block for all of the commands.