Create SendMessage Wrapper, Create StackTraceReporter
The SendMessage wrapper is a static function in Bot. as opposed to the
MessageChannel.sendMessage() provided by JDA, this will automatically
clip messages that are too long into multiple messages of valid length
(2000 chars at this time).

Bot.ReportStackTrace() should be used whenever a command must
abort because of an exception. This will print exception and a
stacktrace to the provided MessageChannel.