Use anonymous classes in favor of functional interfaces.
This organization will place all of a commands relevant information in one place:
 - Command Name
 - Command Desc
 - Command implementation

Instead of the code for a command being implemented as a static function in a different class, the function for the command is now implemented by overriding a abstract method that belongs to the Command enum. Each entry in the Command enum can override this abstract method by providing an anonymous class for each entry. This is done by introducing a scope after the constructor arguments.